Linux port: replace TIOCSETP with <termios.h>.
This is a bit fiddly because in the old 'struct sgttyb' API, all the
terminal flags appear to be smushed together in a single sg_flags
field, but in 'struct termios', they're split up into c_iflag,
c_lflag, c_oflag and c_cflag, so you have to be careful which ones go
where.

Particularly inconvenient was the mode() helper function, which takes
a single pair of int arguments to say what flags it wants to change,
and doesn't say which termios field they live in. Luckily, all current
callers of that function were interested only in flags that live in
c_lflag. Renamed the function to make the limitation clear, so that
someone writing a new call to it won't expect to be able to pass other
flags.