C standards fix: use a dedicated signal handler.
Passing a function to signal() with the wrong prototype has undefined
behaviour (even though it's workable in many cases in practice). So
instead of using the nullary function cleanup() directly as a signal
handler, we should use a dedicated function with exactly the right
prototype, and have that call cleanup().