Fix potential buffer overflow in getpeername().
The 'fromlen' parameter is for input as well as output: on input it
tells the system call how much memory it's _allowed_ to overwrite, and
on output it's adjusted to say how much it actually did write to. So
it's a mistake to pass it in uninitialised.

Also changed 'fromlen' to the approved typedef socklen_t, which is
guaranteed to be the integer type pointer-compatible with that
parameter.