C standards fix: remove trailing junk after #endif.
It's not legal to write '#endif 0'. Putting the 0 in a comment would
be OK, but removing it completely seems just as good, since the #ifs
in question are only a few lines long each, so nobody's at risk of not
being able to work out which #if goes with this #endif.