Fix operator-precedence bug in column calculation.
The intent of this line seems to have been to OR together two batches
of bits from different sources to compute the true column number, then
add 1 to it for the SUPDUP protocol representation. But if you write
'1 + a | b', the + has higher priority, which means that if the column
count had happened to carry from the low bits into the high bits, the
carry would have been lost.

(Pointed out by gcc -Wall.)