C standards fix: char signedness mismatch in sprintf.
nfrontp is an array of unsigned char, so passing it directly to
sprintf triggers a pointer-type mismatch warning. Explicitly cast it
to char *.

Also, to save having to do the same to the subsequent strlen calls,
replace them with just using the return value of sprintf, which
already contains the value you want.